[ddclient] not copy conf file if exist on /etc.

+++ source/applications/ddclient/install.sh
@@ -3,5 +3,5 @@
 cp ddclient /usr/sbin/ddclient
 cp sample-etc_rc.d_init.d_ddclient.ubuntu /etc/init.d/ddclient
 mkdir -p /etc/ddclient/
-cp -f sample-etc_ddclient.conf /etc/ddclient/ddclient.conf
+[ -e /etc/ddclient/ddclient.conf ] || cp -f sample-etc_ddclient.conf /etc/ddclient/ddclient.conf
 mkdir -p /var/cache/ddclient
